requirements.txt
livekit>=0.11.0
livekit-agents>=1.2.0
livekit-plugins-openai>=0.6.0
openai>=1.12.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0

